Refactor orderform component
